- Replaced the bulk of src/benchwrap/cli.py:38-123 with slim Click wrappers that delegate to helper modules while re-exporting the historic symbols so existing tooling keeps working.   - Broke the monolith into focused modules: configuration in src/benchwrap/cli_constants.py:1-17, auth flows in src/benchwrap/cli_auth.py:11-87, progress utilities in src/benchwrap/cli_progress.py:13-   130, benchmark command logic in src/benchwrap/cli_benchmarks.py:19-201, and sync/upload handling in src/benchwrap/cli_sync.py:25-216.   - Introduced injectable subprocess handling in src/benchwrap/cli_benchmarks.py:33-196 so tests can continue monkeypatching benchwrap.cli.subprocess without touching deep internals.
  Testing

  - pytest (fails: tests/test_slurm.py::TestSlurmIntegration::test_slurm_job_submission because sbatch cannot contact a SLURM controller inside the sandbox; all other tests pass).
